+++ bin/belalang/Belalang.h
@@ -1,9 +1,11 @@
-#ifndef BIN_BELALANG_CTX_H_
-#define BIN_BELALANG_CTX_H_
+#ifndef BIN_BELALANG_BELALANG_H_
+#define BIN_BELALANG_BELALANG_H_
 
 #include "llvm/ADT/SmallString.h"
 #include "llvm/ADT/StringRef.h"
 #include "llvm/Support/Error.h"
+#include "llvm/Support/raw_ostream.h"
+#include "muopt/muopt.hpp"
 #include <string>
 
 namespace belalang {
@@ -26,7 +28,18 @@ llvm::Expected<int> link(const BelalangCtx &ctx, llvm::StringRef objectFile,
                          llvm::StringRef executable);
 llvm::Expected<int> execute(llvm::StringRef executable);
 
+int build(muopt::Parser &, const BelalangCtx &);
+int run(muopt::Parser &, const BelalangCtx &);
+int version();
+
+namespace term {
+
+llvm::raw_ostream &error();
+llvm::raw_ostream &warning();
+llvm::raw_ostream &hint();
+
+} // namespace term
 } // namespace cmd
 } // namespace belalang
 
-#endif // BIN_BELALANG_CTX_H_
+#endif // BIN_BELALANG_BELALANG_H_
