+++ bin/belalang/CmdVersion.cpp
@@ -1,4 +1,4 @@
-#include "Cmds.h"
+#include "Belalang.h"
 #include "belalang/Version.h"
 #include "llvm/Support/raw_ostream.h"
